DDD설계 리팩토링
