--- src/main/java/com/highpass/code_review_ide/api/auth/dto/LoginRequest.java
@@ -1,9 +1,0 @@
-package com.highpass.code_review_ide.api.auth.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginRequest(
-        @Email @NotBlank String email,
-        @NotBlank String password
-) {}
