--- src/main/java/com/highpass/code_review_ide/api/auth/dto/LoginResponse.java
@@ -1,6 +1,0 @@
-package com.highpass.code_review_ide.api.auth.dto;
-
-public record LoginResponse(
-        String accessToken,
-        String refreshToken
-) {}
