--- src/main/java/com/highpass/code_review_ide/api/auth/dto/LogoutRequest.java
@@ -1,7 +1,0 @@
-package com.highpass.code_review_ide.api.auth.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record LogoutRequest(
-        @NotBlank String refreshToken
-) {}
