--- src/main/java/com/highpass/code_review_ide/api/auth/dto/MeResponse.java
@@ -1,7 +1,0 @@
-package com.highpass.code_review_ide.api.auth.dto;
-
-public record MeResponse(
-        Long id,
-        String email,
-        String nickname
-) {}
