--- src/main/java/com/highpass/code_review_ide/api/auth/dto/SignupRequest.java
@@ -1,10 +1,0 @@
-package com.highpass.code_review_ide.api.auth.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record SignupRequest(
-        @Email @NotBlank String email,
-        @NotBlank String nickname,
-        @NotBlank String password
-) {}
