--- src/main/java/com/highpass/code_review_ide/api/auth/dto/SignupResponse.java
@@ -1,7 +1,0 @@
-package com.highpass.code_review_ide.api.auth.dto;
-
-public record SignupResponse(
-        Long id,
-        String email,
-        String nickname
-) {}
