--- src/main/java/com/highpass/code_review_ide/api/auth/dto/TokenReissueRequest.java
@@ -1,7 +1,0 @@
-package com.highpass.code_review_ide.api.auth.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record TokenReissueRequest(
-        @NotBlank String refreshToken
-) {}
