--- src/main/java/com/highpass/code_review_ide/api/auth/dto/TokenReissueResponse.java
@@ -1,5 +1,0 @@
-package com.highpass.code_review_ide.api.auth.dto;
-
-public record TokenReissueResponse(
-        String accessToken
-) {}
