+++ src/main/java/com/highpass/code_review_ide/auth/api/dto/request/LoginRequest.java
@@ -1,0 +1,9 @@
+package com.highpass.code_review_ide.auth.api.dto.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginRequest(
+        @Email @NotBlank String email,
+        @NotBlank String password
+) {}
