+++ src/main/java/com/highpass/code_review_ide/auth/api/dto/request/LogoutRequest.java
@@ -1,0 +1,7 @@
+package com.highpass.code_review_ide.auth.api.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LogoutRequest(
+        @NotBlank String refreshToken
+) {}
