+++ src/main/java/com/highpass/code_review_ide/auth/api/dto/request/SignupRequest.java
@@ -1,0 +1,10 @@
+package com.highpass.code_review_ide.auth.api.dto.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record SignupRequest(
+        @Email @NotBlank String email,
+        @NotBlank String nickname,
+        @NotBlank String password
+) {}
