+++ src/main/java/com/highpass/code_review_ide/auth/api/dto/request/TokenReissueRequest.java
@@ -1,0 +1,7 @@
+package com.highpass.code_review_ide.auth.api.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record TokenReissueRequest(
+        @NotBlank String refreshToken
+) {}
