+++ src/main/java/com/highpass/code_review_ide/auth/api/dto/response/LoginResponse.java
@@ -1,0 +1,6 @@
+package com.highpass.code_review_ide.auth.api.dto.response;
+
+public record LoginResponse(
+        String accessToken,
+        String refreshToken
+) {}
