+++ src/main/java/com/highpass/code_review_ide/auth/api/dto/response/MeResponse.java
@@ -1,0 +1,7 @@
+package com.highpass.code_review_ide.auth.api.dto.response;
+
+public record MeResponse(
+        Long id,
+        String email,
+        String nickname
+) {}
