+++ src/main/java/com/highpass/code_review_ide/auth/api/dto/response/SignupResponse.java
@@ -1,0 +1,7 @@
+package com.highpass.code_review_ide.auth.api.dto.response;
+
+public record SignupResponse(
+        Long id,
+        String email,
+        String nickname
+) {}
