+++ src/main/java/com/highpass/code_review_ide/auth/api/dto/response/TokenReissueResponse.java
@@ -1,0 +1,5 @@
+package com.highpass.code_review_ide.auth.api.dto.response;
+
+public record TokenReissueResponse(
+        String accessToken
+) {}
