+++ src/main/java/com/highpass/code_review_ide/auth/domain/email/VerificationPurpose.java
@@ -1,0 +1,7 @@
+package com.highpass.code_review_ide.auth.domain.email;
+
+public enum VerificationPurpose {
+    SIGNUP,
+    FIND_ID,
+    RESET_PASSWORD
+}
