+++ src/main/java/com/highpass/code_review_ide/auth/infrastructure/repository/RefreshTokenRepository.java
@@ -1,0 +1,10 @@
+package com.highpass.code_review_ide.auth.infrastructure.repository;
+
+import com.highpass.code_review_ide.auth.domain.token.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByTokenHash(String tokenHash);
+}
