+++ src/main/java/com/highpass/code_review_ide/auth/infrastructure/security/JwtProvider.java
@@ -1,0 +1,62 @@
+package com.highpass.code_review_ide.auth.infrastructure.security;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+import java.nio.charset.StandardCharsets;
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+import java.util.Date;
+
+@Component
+public class JwtProvider {
+
+    private final SecretKey key;
+    private final int accessMinutes;
+
+    public JwtProvider(
+            @Value("${app.jwt.secret}") String secret,
+            @Value("${app.jwt.access-minutes}") int accessMinutes
+    ) {
+        this.key = Keys.hmacShaKeyFor(secret.getBytes(StandardCharsets.UTF_8));
+        this.accessMinutes = accessMinutes;
+    }
+
+    public String createAccessToken(Long userId, String email) {
+        Instant now = Instant.now();
+        Instant exp = now.plus(accessMinutes, ChronoUnit.MINUTES);
+
+        return Jwts.builder()
+                .subject(String.valueOf(userId))
+                .claim("email", email)
+                .issuedAt(Date.from(now))
+                .expiration(Date.from(exp))
+                .signWith(key)
+                .compact();
+    }
+
+    // 추가: 토큰 검증 + Claims 파싱
+    public Claims parseClaims(String token) {
+        Jws<Claims> jws = Jwts.parser()
+                .verifyWith(key)
+                .build()
+                .parseSignedClaims(token);
+        return jws.getPayload();
+    }
+
+    // 추가: userId 추출
+    public Long getUserId(String token) {
+        return Long.parseLong(parseClaims(token).getSubject());
+    }
+
+    // 추가: email 추출(토큰에 claim으로 넣어뒀으니 활용 가능)
+    public String getEmail(String token) {
+        Object email = parseClaims(token).get("email");
+        return email == null ? null : email.toString();
+    }
+}
