--- src/main/java/com/highpass/code_review_ide/domain/auth/EmailVerification.java
@@ -1,55 +1,0 @@
-package com.highpass.code_review_ide.domain.auth;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "email_verifications")
-@EntityListeners(AuditingEntityListener.class)
-public class EmailVerification {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 255)
-    private String email;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, length = 30)
-    private VerificationPurpose purpose;
-
-    @Column(name = "code_hash", nullable = false, length = 255)
-    private String codeHash;
-
-    @Column(nullable = false)
-    private LocalDateTime expiresAt;
-
-    private LocalDateTime verifiedAt;
-
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    public EmailVerification(
-            String email,
-            VerificationPurpose purpose,
-            String codeHash,
-            LocalDateTime expiresAt
-    ) {
-        this.email = email;
-        this.purpose = purpose;
-        this.codeHash = codeHash;
-        this.expiresAt = expiresAt;
-    }
-
-    public boolean isVerified() {
-        return verifiedAt != null;
-    }
-}
