--- src/main/java/com/highpass/code_review_ide/domain/auth/VerificationPurpose.java
@@ -1,7 +1,0 @@
-package com.highpass.code_review_ide.domain.auth;
-
-public enum VerificationPurpose {
-    SIGNUP,
-    FIND_ID,
-    RESET_PASSWORD
-}
