--- src/main/java/com/highpass/code_review_ide/domain/user/RefreshToken.java
@@ -1,50 +1,0 @@
-package com.highpass.code_review_ide.domain.user;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "refresh_tokens")
-@EntityListeners(AuditingEntityListener.class)
-public class RefreshToken {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @Column(name = "token_hash", nullable = false, length = 255)
-    private String tokenHash;
-
-    @Column(nullable = false)
-    private LocalDateTime expiresAt;
-
-    private LocalDateTime revokedAt;
-
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    public RefreshToken(User user, String tokenHash, LocalDateTime expiresAt) {
-        this.user = user;
-        this.tokenHash = tokenHash;
-        this.expiresAt = expiresAt;
-    }
-
-    public boolean isRevoked() {
-        return revokedAt != null;
-    }
-
-    public void revoke(LocalDateTime now) {
-        this.revokedAt = now;
-    }
-}
