--- src/main/java/com/highpass/code_review_ide/domain/user/RefreshTokenRepository.java
@@ -1,9 +1,0 @@
-package com.highpass.code_review_ide.domain.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByTokenHash(String tokenHash);
-}
