--- src/main/java/com/highpass/code_review_ide/domain/user/UserRepository.java
@@ -1,10 +1,0 @@
-package com.highpass.code_review_ide.domain.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    boolean existsByEmail(String email);
-    boolean existsByNickname(String nickname);
-    Optional<User> findByEmail(String email);
-}
