--- src/main/java/com/highpass/code_review_ide/domain/user/UserStatus.java
@@ -1,6 +1,0 @@
-package com.highpass.code_review_ide.domain.user;
-
-public enum UserStatus {
-    ACTIVE,
-    INACTIVE
-}
