+++ src/main/java/com/highpass/code_review_ide/user/domain/UserStatus.java
@@ -1,0 +1,6 @@
+package com.highpass.code_review_ide.user.domain;
+
+public enum UserStatus {
+    ACTIVE,
+    INACTIVE
+}
