+++ src/main/java/com/highpass/code_review_ide/user/infrastructure/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.highpass.code_review_ide.user.infrastructure.repository;
+
+import com.highpass.code_review_ide.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    boolean existsByEmail(String email);
+    boolean existsByNickname(String nickname);
+    Optional<User> findByEmail(String email);
+}
